Create new packages and refactor
