+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/FollowService.java
@@ -1,0 +1,4 @@
+package edu.byu.cs.tweeter.client.model.service;
+
+public class FollowService {
+}
