+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFeedTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -9,18 +9,19 @@ import java.io.Serializable;
 import java.util.List;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
 import edu.byu.cs.tweeter.util.FakeData;
 import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that retrieves a page of followers.
+ * Background task that retrieves a page of statuses from a user's feed.
  */
-public class GetFollowersTask implements Runnable {
-    private static final String LOG_TAG = "GetFollowersTask";
+public class GetFeedTask implements Runnable {
+    private static final String LOG_TAG = "GetFeedTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String FOLLOWERS_KEY = "followers";
+    public static final String STATUSES_KEY = "statuses";
     public static final String MORE_PAGES_KEY = "more-pages";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
@@ -30,42 +31,42 @@ public class GetFollowersTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * The user whose followers are being retrieved.
+     * The user whose feed is being retrieved.
      * (This can be any user, not just the currently logged-in user.)
      */
     private User targetUser;
     /**
-     * Maximum number of followers to return (i.e., page size).
+     * Maximum number of statuses to return (i.e., page size).
      */
     private int limit;
     /**
-     * The last follower returned in the previous page of results (can be null).
+     * The last status returned in the previous page of results (can be null).
      * This allows the new page to begin where the previous page ended.
      */
-    private User lastFollower;
+    private Status lastStatus;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public GetFollowersTask(AuthToken authToken, User targetUser, int limit, User lastFollower,
-                            Handler messageHandler) {
+    public GetFeedTask(AuthToken authToken, User targetUser, int limit, Status lastStatus,
+                       Handler messageHandler) {
         this.authToken = authToken;
         this.targetUser = targetUser;
         this.limit = limit;
-        this.lastFollower = lastFollower;
+        this.lastStatus = lastStatus;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
-            Pair<List<User>, Boolean> pageOfUsers = getFollowers();
+            Pair<List<Status>, Boolean> pageOfStatus = getFeed();
 
-            List<User> followers = pageOfUsers.getFirst();
-            boolean hasMorePages = pageOfUsers.getSecond();
+            List<Status> statuses = pageOfStatus.getFirst();
+            boolean hasMorePages = pageOfStatus.getSecond();
 
-            sendSuccessMessage(followers, hasMorePages);
+            sendSuccessMessage(statuses, hasMorePages);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -77,15 +78,15 @@ public class GetFollowersTask implements Runnable {
         return new FakeData();
     }
 
-    private Pair<List<User>, Boolean> getFollowers() {
-        Pair<List<User>, Boolean> pageOfUsers = getFakeData().getPageOfUsers(lastFollower, limit, targetUser);
-        return pageOfUsers;
+    private Pair<List<Status>, Boolean> getFeed() {
+        Pair<List<Status>, Boolean> pageOfStatus = getFakeData().getPageOfStatus(lastStatus, limit);
+        return pageOfStatus;
     }
 
-    private void sendSuccessMessage(List<User> followers, boolean hasMorePages) {
+    private void sendSuccessMessage(List<Status> statuses, boolean hasMorePages) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(FOLLOWERS_KEY, (Serializable) followers);
+        msgBundle.putSerializable(STATUSES_KEY, (Serializable) statuses);
         msgBundle.putBoolean(MORE_PAGES_KEY, hasMorePages);
 
         Message msg = Message.obtain();
@@ -115,5 +116,4 @@ public class GetFollowersTask implements Runnable {
 
         messageHandler.sendMessage(msg);
     }
-
 }
