+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowersCountTask.java
@@ -1,28 +1,21 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 import android.util.Log;
 
-import java.io.Serializable;
-import java.util.List;
-
 import edu.byu.cs.tweeter.model.domain.AuthToken;
-import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
-import edu.byu.cs.tweeter.util.FakeData;
-import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that retrieves a page of statuses from a user's story.
+ * Background task that queries how many followers a user has.
  */
-public class GetStoryTask implements Runnable {
-    private static final String LOG_TAG = "GetStoryTask";
+public class GetFollowersCountTask implements Runnable {
+    private static final String LOG_TAG = "GetFollowersCountTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String STATUSES_KEY = "statuses";
-    public static final String MORE_PAGES_KEY = "more-pages";
+    public static final String COUNT_KEY = "count";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
@@ -31,42 +24,26 @@ public class GetStoryTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * The user whose story is being retrieved.
+     * The user whose follower count is being retrieved.
      * (This can be any user, not just the currently logged-in user.)
      */
     private User targetUser;
-    /**
-     * Maximum number of statuses to return (i.e., page size).
-     */
-    private int limit;
-    /**
-     * The last status returned in the previous page of results (can be null).
-     * This allows the new page to begin where the previous page ended.
-     */
-    private Status lastStatus;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public GetStoryTask(AuthToken authToken, User targetUser, int limit, Status lastStatus,
-                        Handler messageHandler) {
+    public GetFollowersCountTask(AuthToken authToken, User targetUser, Handler messageHandler) {
         this.authToken = authToken;
         this.targetUser = targetUser;
-        this.limit = limit;
-        this.lastStatus = lastStatus;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
-            Pair<List<Status>, Boolean> pageOfStatus = getStory();
 
-            List<Status> statuses = pageOfStatus.getFirst();
-            boolean hasMorePages = pageOfStatus.getSecond();
-
-            sendSuccessMessage(statuses, hasMorePages);
+            sendSuccessMessage(20);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -74,20 +51,10 @@ public class GetStoryTask implements Runnable {
         }
     }
 
-    private FakeData getFakeData() {
-        return new FakeData();
-    }
-
-    private Pair<List<Status>, Boolean> getStory() {
-        Pair<List<Status>, Boolean> pageOfStatus = getFakeData().getPageOfStatus(lastStatus, limit);
-        return pageOfStatus;
-    }
-
-    private void sendSuccessMessage(List<Status> statuses, boolean hasMorePages) {
+    private void sendSuccessMessage(int count) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(STATUSES_KEY, (Serializable) statuses);
-        msgBundle.putBoolean(MORE_PAGES_KEY, hasMorePages);
+        msgBundle.putInt(COUNT_KEY, count);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
