+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowersTask.java
@@ -1,21 +1,27 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 import android.util.Log;
 
+import java.io.Serializable;
+import java.util.List;
+
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.util.FakeData;
+import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that queries how many other users a specified user is following.
+ * Background task that retrieves a page of followers.
  */
-public class GetFollowingCountTask implements Runnable {
-    private static final String LOG_TAG = "GetFollowingCountTask";
+public class GetFollowersTask implements Runnable {
+    private static final String LOG_TAG = "GetFollowersTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String COUNT_KEY = "count";
+    public static final String FOLLOWERS_KEY = "followers";
+    public static final String MORE_PAGES_KEY = "more-pages";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
@@ -24,26 +30,42 @@ public class GetFollowingCountTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * The user whose following count is being retrieved.
+     * The user whose followers are being retrieved.
      * (This can be any user, not just the currently logged-in user.)
      */
     private User targetUser;
+    /**
+     * Maximum number of followers to return (i.e., page size).
+     */
+    private int limit;
+    /**
+     * The last follower returned in the previous page of results (can be null).
+     * This allows the new page to begin where the previous page ended.
+     */
+    private User lastFollower;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public GetFollowingCountTask(AuthToken authToken, User targetUser, Handler messageHandler) {
+    public GetFollowersTask(AuthToken authToken, User targetUser, int limit, User lastFollower,
+                            Handler messageHandler) {
         this.authToken = authToken;
         this.targetUser = targetUser;
+        this.limit = limit;
+        this.lastFollower = lastFollower;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
+            Pair<List<User>, Boolean> pageOfUsers = getFollowers();
 
-            sendSuccessMessage(20);
+            List<User> followers = pageOfUsers.getFirst();
+            boolean hasMorePages = pageOfUsers.getSecond();
+
+            sendSuccessMessage(followers, hasMorePages);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -51,10 +73,20 @@ public class GetFollowingCountTask implements Runnable {
         }
     }
 
-    private void sendSuccessMessage(int count) {
+    private FakeData getFakeData() {
+        return new FakeData();
+    }
+
+    private Pair<List<User>, Boolean> getFollowers() {
+        Pair<List<User>, Boolean> pageOfUsers = getFakeData().getPageOfUsers(lastFollower, limit, targetUser);
+        return pageOfUsers;
+    }
+
+    private void sendSuccessMessage(List<User> followers, boolean hasMorePages) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putInt(COUNT_KEY, count);
+        msgBundle.putSerializable(FOLLOWERS_KEY, (Serializable) followers);
+        msgBundle.putBoolean(MORE_PAGES_KEY, hasMorePages);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
@@ -83,4 +115,5 @@ public class GetFollowingCountTask implements Runnable {
 
         messageHandler.sendMessage(msg);
     }
+
 }
