+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowingCountTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -7,50 +7,43 @@ import android.util.Log;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
-import edu.byu.cs.tweeter.util.FakeData;
-import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that logs in a user (i.e., starts a session).
+ * Background task that queries how many other users a specified user is following.
  */
-public class LoginTask implements Runnable {
-
-    private static final String LOG_TAG = "LoginTask";
+public class GetFollowingCountTask implements Runnable {
+    private static final String LOG_TAG = "GetFollowingCountTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String USER_KEY = "user";
-    public static final String AUTH_TOKEN_KEY = "auth-token";
+    public static final String COUNT_KEY = "count";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
     /**
-     * The user's username (or "alias" or "handle"). E.g., "@susan".
+     * Auth token for logged-in user.
      */
-    private String username;
+    private AuthToken authToken;
     /**
-     * The user's password.
+     * The user whose following count is being retrieved.
+     * (This can be any user, not just the currently logged-in user.)
      */
-    private String password;
+    private User targetUser;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public LoginTask(String username, String password, Handler messageHandler) {
-        this.username = username;
-        this.password = password;
+    public GetFollowingCountTask(AuthToken authToken, User targetUser, Handler messageHandler) {
+        this.authToken = authToken;
+        this.targetUser = targetUser;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
-            Pair<User, AuthToken> loginResult = doLogin();
 
-            User loggedInUser = loginResult.getFirst();
-            AuthToken authToken = loginResult.getSecond();
-
-            sendSuccessMessage(loggedInUser, authToken);
+            sendSuccessMessage(20);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -58,21 +51,10 @@ public class LoginTask implements Runnable {
         }
     }
 
-    private FakeData getFakeData() {
-        return new FakeData();
-    }
-
-    private Pair<User, AuthToken> doLogin() {
-        User loggedInUser = getFakeData().getFirstUser();
-        AuthToken authToken = getFakeData().getAuthToken();
-        return new Pair<>(loggedInUser, authToken);
-    }
-
-    private void sendSuccessMessage(User loggedInUser, AuthToken authToken) {
+    private void sendSuccessMessage(int count) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(USER_KEY, loggedInUser);
-        msgBundle.putSerializable(AUTH_TOKEN_KEY, authToken);
+        msgBundle.putInt(COUNT_KEY, count);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
