+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowingTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -9,19 +9,18 @@ import java.io.Serializable;
 import java.util.List;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
-import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
 import edu.byu.cs.tweeter.util.FakeData;
 import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that retrieves a page of statuses from a user's feed.
+ * Background task that retrieves a page of other users being followed by a specified user.
  */
-public class GetFeedTask implements Runnable {
-    private static final String LOG_TAG = "GetFeedTask";
+public class GetFollowingTask implements Runnable {
+    private static final String LOG_TAG = "GetFollowingTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String STATUSES_KEY = "statuses";
+    public static final String FOLLOWEES_KEY = "followees";
     public static final String MORE_PAGES_KEY = "more-pages";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
@@ -31,45 +30,45 @@ public class GetFeedTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * The user whose feed is being retrieved.
+     * The user whose following is being retrieved.
      * (This can be any user, not just the currently logged-in user.)
      */
     private User targetUser;
     /**
-     * Maximum number of statuses to return (i.e., page size).
+     * Maximum number of followed users to return (i.e., page size).
      */
     private int limit;
     /**
-     * The last status returned in the previous page of results (can be null).
+     * The last person being followed returned in the previous page of results (can be null).
      * This allows the new page to begin where the previous page ended.
      */
-    private Status lastStatus;
+    private User lastFollowee;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public GetFeedTask(AuthToken authToken, User targetUser, int limit, Status lastStatus,
-                       Handler messageHandler) {
+    public GetFollowingTask(AuthToken authToken, User targetUser, int limit, User lastFollowee,
+                            Handler messageHandler) {
         this.authToken = authToken;
         this.targetUser = targetUser;
         this.limit = limit;
-        this.lastStatus = lastStatus;
+        this.lastFollowee = lastFollowee;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
-            Pair<List<Status>, Boolean> pageOfStatus = getFeed();
+            Pair<List<User>, Boolean> pageOfUsers = getFollowees();
 
-            List<Status> statuses = pageOfStatus.getFirst();
-            boolean hasMorePages = pageOfStatus.getSecond();
+            List<User> followees = pageOfUsers.getFirst();
+            boolean hasMorePages = pageOfUsers.getSecond();
 
-            sendSuccessMessage(statuses, hasMorePages);
+            sendSuccessMessage(followees, hasMorePages);
 
         } catch (Exception ex) {
-            Log.e(LOG_TAG, ex.getMessage(), ex);
+            Log.e(LOG_TAG, "Failed to get followees", ex);
             sendExceptionMessage(ex);
         }
     }
@@ -78,15 +77,15 @@ public class GetFeedTask implements Runnable {
         return new FakeData();
     }
 
-    private Pair<List<Status>, Boolean> getFeed() {
-        Pair<List<Status>, Boolean> pageOfStatus = getFakeData().getPageOfStatus(lastStatus, limit);
-        return pageOfStatus;
+    private Pair<List<User>, Boolean> getFollowees() {
+        return getFakeData().getPageOfUsers((User) lastFollowee, limit, targetUser);
     }
 
-    private void sendSuccessMessage(List<Status> statuses, boolean hasMorePages) {
+
+    private void sendSuccessMessage(List<User> followees, boolean hasMorePages) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(STATUSES_KEY, (Serializable) statuses);
+        msgBundle.putSerializable(FOLLOWEES_KEY, (Serializable) followees);
         msgBundle.putBoolean(MORE_PAGES_KEY, hasMorePages);
 
         Message msg = Message.obtain();
@@ -116,4 +115,5 @@ public class GetFeedTask implements Runnable {
 
         messageHandler.sendMessage(msg);
     }
+
 }
