+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetUserTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -6,15 +6,17 @@ import android.os.Message;
 import android.util.Log;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
-import edu.byu.cs.tweeter.model.domain.Status;
+import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.util.FakeData;
 
 /**
- * Background task that posts a new status sent by a user.
+ * Background task that returns the profile for a specified user.
  */
-public class PostStatusTask implements Runnable {
-    private static final String LOG_TAG = "PostStatusTask";
+public class GetUserTask implements Runnable {
+    private static final String LOG_TAG = "GetUserTask";
 
     public static final String SUCCESS_KEY = "success";
+    public static final String USER_KEY = "user";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
@@ -23,26 +25,26 @@ public class PostStatusTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * The new status being sent. Contains all properties of the status,
-     * including the identity of the user sending the status.
+     * Alias (or handle) for user whose profile is being retrieved.
      */
-    private Status status;
+    private String alias;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public PostStatusTask(AuthToken authToken, Status status, Handler messageHandler) {
+    public GetUserTask(AuthToken authToken, String alias, Handler messageHandler) {
         this.authToken = authToken;
-        this.status = status;
+        this.alias = alias;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
+            User user = getUser();
 
-            sendSuccessMessage();
+            sendSuccessMessage(user);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -50,9 +52,19 @@ public class PostStatusTask implements Runnable {
         }
     }
 
-    private void sendSuccessMessage() {
+    private FakeData getFakeData() {
+        return new FakeData();
+    }
+
+    private User getUser() {
+        User user = getFakeData().findUserByAlias(alias);
+        return user;
+    }
+
+    private void sendSuccessMessage(User user) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
+        msgBundle.putSerializable(USER_KEY, user);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
