+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/IsFollowerTask.java
@@ -1,22 +1,23 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 import android.util.Log;
 
+import java.util.Random;
+
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
-import edu.byu.cs.tweeter.util.FakeData;
 
 /**
- * Background task that returns the profile for a specified user.
+ * Background task that determines if one user is following another.
  */
-public class GetUserTask implements Runnable {
-    private static final String LOG_TAG = "GetUserTask";
+public class IsFollowerTask implements Runnable {
+    private static final String LOG_TAG = "IsFollowerTask";
 
     public static final String SUCCESS_KEY = "success";
-    public static final String USER_KEY = "user";
+    public static final String IS_FOLLOWER_KEY = "is-follower";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
@@ -25,26 +26,30 @@ public class GetUserTask implements Runnable {
      */
     private AuthToken authToken;
     /**
-     * Alias (or handle) for user whose profile is being retrieved.
+     * The alleged follower.
      */
-    private String alias;
+    private User follower;
+    /**
+     * The alleged followee.
+     */
+    private User followee;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public GetUserTask(AuthToken authToken, String alias, Handler messageHandler) {
+    public IsFollowerTask(AuthToken authToken, User follower, User followee, Handler messageHandler) {
         this.authToken = authToken;
-        this.alias = alias;
+        this.follower = follower;
+        this.followee = followee;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
-            User user = getUser();
 
-            sendSuccessMessage(user);
+            sendSuccessMessage(new Random().nextInt() > 0);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -52,19 +57,10 @@ public class GetUserTask implements Runnable {
         }
     }
 
-    private FakeData getFakeData() {
-        return new FakeData();
-    }
-
-    private User getUser() {
-        User user = getFakeData().findUserByAlias(alias);
-        return user;
-    }
-
-    private void sendSuccessMessage(User user) {
+    private void sendSuccessMessage(boolean isFollower) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
-        msgBundle.putSerializable(USER_KEY, user);
+        msgBundle.putBoolean(IS_FOLLOWER_KEY, isFollower);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
