+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/LoginTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -6,36 +6,51 @@ import android.os.Message;
 import android.util.Log;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.util.FakeData;
+import edu.byu.cs.tweeter.util.Pair;
 
 /**
- * Background task that logs out a user (i.e., ends a session).
+ * Background task that logs in a user (i.e., starts a session).
  */
-public class LogoutTask implements Runnable {
-    private static final String LOG_TAG = "LogoutTask";
+public class LoginTask implements Runnable {
+
+    private static final String LOG_TAG = "LoginTask";
 
     public static final String SUCCESS_KEY = "success";
+    public static final String USER_KEY = "user";
+    public static final String AUTH_TOKEN_KEY = "auth-token";
     public static final String MESSAGE_KEY = "message";
     public static final String EXCEPTION_KEY = "exception";
 
     /**
-     * Auth token for logged-in user.
+     * The user's username (or "alias" or "handle"). E.g., "@susan".
      */
-    private AuthToken authToken;
+    private String username;
+    /**
+     * The user's password.
+     */
+    private String password;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public LogoutTask(AuthToken authToken, Handler messageHandler) {
-        this.authToken = authToken;
+    public LoginTask(String username, String password, Handler messageHandler) {
+        this.username = username;
+        this.password = password;
         this.messageHandler = messageHandler;
     }
 
     @Override
     public void run() {
         try {
+            Pair<User, AuthToken> loginResult = doLogin();
 
-            sendSuccessMessage();
+            User loggedInUser = loginResult.getFirst();
+            AuthToken authToken = loginResult.getSecond();
+
+            sendSuccessMessage(loggedInUser, authToken);
 
         } catch (Exception ex) {
             Log.e(LOG_TAG, ex.getMessage(), ex);
@@ -43,9 +58,21 @@ public class LogoutTask implements Runnable {
         }
     }
 
-    private void sendSuccessMessage() {
+    private FakeData getFakeData() {
+        return new FakeData();
+    }
+
+    private Pair<User, AuthToken> doLogin() {
+        User loggedInUser = getFakeData().getFirstUser();
+        AuthToken authToken = getFakeData().getAuthToken();
+        return new Pair<>(loggedInUser, authToken);
+    }
+
+    private void sendSuccessMessage(User loggedInUser, AuthToken authToken) {
         Bundle msgBundle = new Bundle();
         msgBundle.putBoolean(SUCCESS_KEY, true);
+        msgBundle.putSerializable(USER_KEY, loggedInUser);
+        msgBundle.putSerializable(AUTH_TOKEN_KEY, authToken);
 
         Message msg = Message.obtain();
         msg.setData(msgBundle);
