+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/PostStatusTask.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.backgroundTask;
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
 
 import android.os.Bundle;
 import android.os.Handler;
@@ -6,13 +6,13 @@ import android.os.Message;
 import android.util.Log;
 
 import edu.byu.cs.tweeter.model.domain.AuthToken;
-import edu.byu.cs.tweeter.model.domain.User;
+import edu.byu.cs.tweeter.model.domain.Status;
 
 /**
- * Background task that establishes a following relationship between two users.
+ * Background task that posts a new status sent by a user.
  */
-public class FollowTask implements Runnable {
-    private static final String LOG_TAG = "FollowTask";
+public class PostStatusTask implements Runnable {
+    private static final String LOG_TAG = "PostStatusTask";
 
     public static final String SUCCESS_KEY = "success";
     public static final String MESSAGE_KEY = "message";
@@ -20,21 +20,21 @@ public class FollowTask implements Runnable {
 
     /**
      * Auth token for logged-in user.
-     * This user is the "follower" in the relationship.
      */
     private AuthToken authToken;
     /**
-     * The user that is being followed.
+     * The new status being sent. Contains all properties of the status,
+     * including the identity of the user sending the status.
      */
-    private User followee;
+    private Status status;
     /**
      * Message handler that will receive task results.
      */
     private Handler messageHandler;
 
-    public FollowTask(AuthToken authToken, User followee, Handler messageHandler) {
+    public PostStatusTask(AuthToken authToken, Status status, Handler messageHandler) {
         this.authToken = authToken;
-        this.followee = followee;
+        this.status = status;
         this.messageHandler = messageHandler;
     }
