+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/UnfollowTask.java
@@ -1,0 +1,84 @@
+package edu.byu.cs.tweeter.client.model.service.backgroundTask;
+
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
+import android.util.Log;
+
+import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.User;
+
+/**
+ * Background task that removes a following relationship between two users.
+ */
+public class UnfollowTask implements Runnable {
+    private static final String LOG_TAG = "UnfollowTask";
+
+    public static final String SUCCESS_KEY = "success";
+    public static final String MESSAGE_KEY = "message";
+    public static final String EXCEPTION_KEY = "exception";
+
+    /**
+     * Auth token for logged-in user.
+     * This user is the "follower" in the relationship.
+     */
+    private AuthToken authToken;
+    /**
+     * The user that is being followed.
+     */
+    private User followee;
+    /**
+     * Message handler that will receive task results.
+     */
+    private Handler messageHandler;
+
+    public UnfollowTask(AuthToken authToken, User followee, Handler messageHandler) {
+        this.authToken = authToken;
+        this.followee = followee;
+        this.messageHandler = messageHandler;
+    }
+
+    @Override
+    public void run() {
+        try {
+
+            sendSuccessMessage();
+
+        } catch (Exception ex) {
+            Log.e(LOG_TAG, ex.getMessage(), ex);
+            sendExceptionMessage(ex);
+        }
+    }
+
+    private void sendSuccessMessage() {
+        Bundle msgBundle = new Bundle();
+        msgBundle.putBoolean(SUCCESS_KEY, true);
+
+        Message msg = Message.obtain();
+        msg.setData(msgBundle);
+
+        messageHandler.sendMessage(msg);
+    }
+
+    private void sendFailedMessage(String message) {
+        Bundle msgBundle = new Bundle();
+        msgBundle.putBoolean(SUCCESS_KEY, false);
+        msgBundle.putString(MESSAGE_KEY, message);
+
+        Message msg = Message.obtain();
+        msg.setData(msgBundle);
+
+        messageHandler.sendMessage(msg);
+    }
+
+    private void sendExceptionMessage(Exception exception) {
+        Bundle msgBundle = new Bundle();
+        msgBundle.putBoolean(SUCCESS_KEY, false);
+        msgBundle.putSerializable(EXCEPTION_KEY, exception);
+
+        Message msg = Message.obtain();
+        msg.setData(msgBundle);
+
+        messageHandler.sendMessage(msg);
+    }
+}
